:zap: aggregate test
